backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai>=1.0.0
anthropic==0.25.0
httpx==0.24.1
google-cloud-aiplatform>=1.40.0
python-docx==1.1.0
python-pptx==0.6.23
reportlab==4.0.7
google-auth==2.25.2
Pillow==10.1.0
numpy==1.26.2
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1

# OpenTelemetry Core - using compatible versions
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.42b0
opentelemetry-exporter-otlp>=1.20.0

# OpenTelemetry Instrumentation
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-instrumentation-aiohttp-client>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-instrumentation-sqlite3>=0.42b0
opentelemetry-instrumentation-logging>=0.42b0